Adds failing parser ordering test
When parsing in one order we get an `ok` tuple, but in the other order
we get an `error` tuple. I believe this is a mistake, but I'm not sure
how to fix the issue. I wrote a very simple test to show what is
happening.

Amos King @adkron <amos@binarynoggin.com>